Shred `FusionUtils` grab-bag into more sensible parts.
