+++ runtime/src/main/java/dev/ionfusion/fusion/FusionValue.java
@@ -4,8 +4,8 @@
 package dev.ionfusion.fusion;
 
 import static dev.ionfusion.fusion.FusionIo.makeIonizeError;
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_STRING_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_STRING_ARRAY;
 
 import com.amazon.ion.IonValue;
 import com.amazon.ion.ValueFactory;
