+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxKeyword.java
@@ -4,7 +4,7 @@
 package dev.ionfusion.fusion;
 
 import static dev.ionfusion.fusion.SyntaxException.makeSyntaxError;
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_OBJECT_ARRAY;
 
 import com.amazon.ion.IonException;
 import com.amazon.ion.IonWriter;
