+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxString.java
@@ -3,7 +3,7 @@
 
 package dev.ionfusion.fusion;
 
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_OBJECT_ARRAY;
 
 import dev.ionfusion.fusion.FusionString.BaseString;
 import dev.ionfusion.runtime.base.SourceLocation;
