+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxValue.java
@@ -6,7 +6,7 @@ package dev.ionfusion.fusion;
 import static dev.ionfusion.fusion.FusionCompare.isSame;
 import static dev.ionfusion.fusion.FusionSexp.emptySexp;
 import static dev.ionfusion.fusion.FusionSexp.pair;
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_OBJECT_ARRAY;
 import static java.lang.Boolean.TRUE;
 
 import com.amazon.ion.IonValue;
