+++ runtime/src/test/java/dev/ionfusion/fusion/LobTestCase.java
@@ -8,8 +8,8 @@ import static dev.ionfusion.fusion.FusionLob.unsafeLobBytesNoCopy;
 import static dev.ionfusion.fusion.FusionString.unsafeStringToJavaString;
 import static dev.ionfusion.fusion.FusionValue.annotate;
 import static dev.ionfusion.fusion.FusionValue.isAnyNull;
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_BYTE_ARRAY;
-import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_STRING_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_BYTE_ARRAY;
+import static dev.ionfusion.runtime._private.util.Empties.EMPTY_STRING_ARRAY;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
